+++ src/voxel/voxel.rs
@@ -2,7 +2,7 @@ use block_mesh::{MergeVoxel, Voxel as MeshableVoxel};
 use serde::{Serialize, Deserialize};
 
 #[derive(Clone, Copy, Hash, Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub struct Voxel(pub u8);
+pub struct Voxel(pub u16);
 
 impl Voxel {
     pub const EMPTY_VOXEL: Self = Self(0);
@@ -26,7 +26,7 @@ impl MeshableVoxel for Voxel {
 }
 
 impl MergeVoxel for Voxel {
-    type MergeValue = u8;
+    type MergeValue = u16;
 
     #[inline]
     fn merge_value(&self) -> Self::MergeValue {
@@ -35,11 +35,11 @@ impl MergeVoxel for Voxel {
 }
 
 pub trait MaterialVoxel: MergeVoxel + MeshableVoxel {
-    fn as_mat_id(&self) -> u8;
+    fn as_mat_id(&self) -> u16;
 }
 
 impl MaterialVoxel for Voxel {
-    fn as_mat_id(&self) -> u8 {
+    fn as_mat_id(&self) -> u16 {
         self.0
     }
 }
